1240-refactor for class renderer
prep for future functionality work
