+++ src/diagrams/class/svgDraw.spec.js
@@ -1,11 +1,11 @@
 /* eslint-env jasmine */
-import memberRenderer from './classMemberRenderer';
+import svgDraw from './svgDraw';
 
 describe('class member Renderer, ', function () {
   describe('when parsing text to build method display string', function () {
     it('should handle simple method declaration', function () {
       const str = 'foo()';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo()');
       expect(actual.cssStyle).toBe('');
@@ -13,7 +13,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle public visibility', function () {
       const str = '+foo()';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('+foo()');
       expect(actual.cssStyle).toBe('');
@@ -21,7 +21,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle private visibility', function () {
       const str = '-foo()';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('-foo()');
       expect(actual.cssStyle).toBe('');
@@ -29,7 +29,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle protected visibility', function () {
       const str = '#foo()';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('#foo()');
       expect(actual.cssStyle).toBe('');
@@ -37,7 +37,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle package/internal visibility', function () {
       const str = '~foo()';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('~foo()');
       expect(actual.cssStyle).toBe('');
@@ -45,7 +45,7 @@ describe('class member Renderer, ', function () {
 
     it('should ignore unknown character for visibility', function () {
       const str = '!foo()';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo()');
       expect(actual.cssStyle).toBe('');
@@ -53,7 +53,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle abstract classifier', function () {
       const str = 'foo()*';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo()');
       expect(actual.cssStyle).toBe('font-style:italic;');
@@ -61,7 +61,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle static classifier', function () {
       const str = 'foo()$';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo()');
       expect(actual.cssStyle).toBe('text-decoration:underline;');
@@ -69,7 +69,7 @@ describe('class member Renderer, ', function () {
 
     it('should ignore unknown character for classifier', function () {
       const str = 'foo()!';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo()');
       expect(actual.cssStyle).toBe('');
@@ -77,7 +77,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle simple method declaration with parameters', function () {
       const str = 'foo(int id)';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
       
       expect(actual.displayText).toBe('foo(int id)');
       expect(actual.cssStyle).toBe('');
@@ -85,7 +85,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle simple method declaration with single item in parameters', function () {
       const str = 'foo(id)';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo(id)');
       expect(actual.cssStyle).toBe('');
@@ -93,7 +93,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle simple method declaration with single item in parameters with extra spaces', function () {
       const str = ' foo ( id) ';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo(id)');
       expect(actual.cssStyle).toBe('');
@@ -101,7 +101,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle method declaration with return value', function () {
       const str = 'foo(id) int';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo(id) : int');
       expect(actual.cssStyle).toBe('');
@@ -109,7 +109,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle method declaration with generic return value', function () {
       const str = 'foo(id) List~int~';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo(id) : List<int>');
       expect(actual.cssStyle).toBe('');
@@ -117,7 +117,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle method declaration with generic parameter', function () {
       const str = 'foo(List~int~)';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('foo(List<int>)');
       expect(actual.cssStyle).toBe('');
@@ -125,33 +125,17 @@ describe('class member Renderer, ', function () {
 
     it('should handle method declaration with all possible markup', function () {
       const str = '+foo (  List~int~ ids  )* List~Item~';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('+foo(List<int> ids) : List<Item>');
       expect(actual.cssStyle).toBe('font-style:italic;');
     });
   });
 
-  describe('when parsing text for generic types', function () {
-    it('should handle open and close brackets in correct order', function () {
-      const str = 'foo(List~Item~)';
-      let actual = memberRenderer.parseGenericTypes(str);
-
-      expect(actual).toBe('foo(List<Item>)');
-    });
-
-    it('should handle open and close brackets in correct order with multiple usages', function () {
-      const str = 'foo(List~Item~) List~Item~';
-      let actual = memberRenderer.parseGenericTypes(str);
-
-      expect(actual).toBe('foo(List<Item>) List<Item>');
-    });
-  });
-
   describe('when parsing text to build field display string', function () {
     it('should handle simple field declaration', function () {
       const str = 'int[] ids';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('int[] ids');
       expect(actual.cssStyle).toBe('');
@@ -159,7 +143,7 @@ describe('class member Renderer, ', function () {
 
     it('should handle field declaration with generic type', function () {
       const str = 'List~int~ ids';
-      let actual = memberRenderer.parseMember(str);
+      let actual = svgDraw.parseMember(str);
 
       expect(actual.displayText).toBe('List<int> ids');
       expect(actual.cssStyle).toBe('');
